+++ frontend/src/pages/modules/PeakProphetModule.tsx
@@ -460,6 +460,7 @@ const PeakProphetModule: React.FC = () => {
                 </div>
               </div>
 
+
               <button
                 type="submit"
                 disabled={loading}